removed dead code, styled slider and deleted unused component
